# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  dto.Avatar:
    properties:
      id:
        type: string
      image:
        type: string
      private:
        type: boolean
    type: object
  dto.Bot:
    properties:
      attack:
        type: integer
      avatar:
        type: string
      createdAt:
        type: string
      defense:
        type: integer
      hp:
        type: integer
      id:
        type: string
      level:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.User:
    properties:
      armsEquipmentItemId:
        type: string
      attack:
        type: integer
      avatar:
        $ref: '#/definitions/dto.Avatar'
      beltEquipmentItemId:
        type: string
      chestEquipmentItemId:
        type: string
      createdAt:
        type: string
      currentHp:
        type: integer
      defense:
        type: integer
      email:
        type: string
      exp:
        type: integer
      feetEquipmentItemId:
        type: string
      freeStats:
        type: integer
      gold:
        type: integer
      handsEquipmentItemId:
        type: string
      headEquipmentItemId:
        type: string
      hp:
        type: integer
      id:
        type: string
      inFight:
        type: boolean
      legsEquipmentItemId:
        type: string
      level:
        type: integer
      locationSlug:
        type: string
      neckEquipmentItemId:
        type: string
      ring1EquipmentItemId:
        type: string
      ring2EquipmentItemId:
        type: string
      ring3EquipmentItemId:
        type: string
      ring4EquipmentItemId:
        type: string
      shieldEquipmentItemId:
        type: string
      username:
        type: string
      weaponEquipmentItemId:
        type: string
    type: object
  handlers.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.User'
    type: object
  handlers.BotResponse:
    properties:
      bots:
        items:
          $ref: '#/definitions/dto.Bot'
        type: array
    type: object
...
  handlers.SignInRequest:
    properties:
      password:
        example: password
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handlers.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Game API for Moonshine
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Moonshine API
  version: "1.0"
paths:
  /api/auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Sign in request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign in
      tags:
      - auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Sign up request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /api/bots/{location_slug}:
    get:
      consumes:
      - application/json
      description: Get list of bots in a specific location
      parameters:
      - description: Location slug
        in: path
        name: location_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BotResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get bots by location
      tags:
      - bots
  /api/bots/{slug}/attack:
    post:
      consumes:
      - application/json
      description: Initiate a fight with a bot
      parameters:
      - description: Bot slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Attack a bot
      tags:
      - bots
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"